Revert "Merge pull request #1 from DolphinGui/ctng-actions"
This reverts commit a6185dd96c977afe622f3b7ceda243792be01d1f, reversing
changes made to ec980d8e5bb63db66a85007ad29334509c8f8a3c.
As it turns out crosstool-NG is really slow when cross-compiling, since
it rebuilds the native toolchain. Also, it doesn't support macOS, which
I found a way to hack. There may be changes in the future to ctng that
changes this, but for now I'm just going back to shell scripts